provider: fix Awareness change handlers not called upon provider destroy (#217)
* fix: Awareness change handlers not called upon provider destroy

* test

+++ tests/provider/onAwarenessChange.js
@@ -13,6 +13,7 @@ context('provider/onAwarenessChange', () => {
 
     server.configure({ port: 4000 }).listen()
 
+    let called = false
     const client = new HocuspocusProvider({
       url: 'ws://127.0.0.1:4000',
       name: 'hocuspocus-test',
@@ -22,12 +23,38 @@ context('provider/onAwarenessChange', () => {
         client.setAwarenessField('foo', 'bar')
       },
       onAwarenessChange: ({ states }) => {
+        if (called) return
+        called = true
+
         server.destroy()
         client.destroy()
 
         assert.strictEqual(states.length, 1)
         assert.strictEqual(states[0].foo, 'bar')
 
+        done()
+      },
+    })
+  })
+
+  it('onAwarenessChange callback is executed on provider destroy', done => {
+    const server = new Hocuspocus()
+
+    server.configure({ port: 4000 }).listen()
+
+    const client = new HocuspocusProvider({
+      url: 'ws://127.0.0.1:4000',
+      name: 'hocuspocus-test',
+      document: ydoc,
+      WebSocketPolyfill: WebSocket,
+      maxAttempts: 1,
+      onConnect: () => {
+        client.destroy()
+      },
+      onAwarenessChange: ({ states }) => {
+        server.destroy()
+
+        assert.strictEqual(states.length, 0)
         done()
       },
     })
@@ -77,8 +104,8 @@ context('provider/onAwarenessChange', () => {
           },
         ])
 
-        client.destroy()
         server.destroy()
+        client.destroy()
 
         done()
       },
@@ -107,8 +134,9 @@ context('provider/onAwarenessChange', () => {
           assert.strictEqual(player2, true)
 
           server.destroy()
-          client.destroy()
           anotherClient.destroy()
+          client.destroy()
+
           done()
         }
       },
